Move more utils to RESTAPI and project ProjectAPI
